+++ packages/remix-netlify-edge/remix-deno/README.md
@@ -1,0 +1,4 @@
+# @remix-run/deno
+
+`@remix-run/deno` package is temporarily inlined within this directory while Deno support is experimental.
+In the future, this directory would be removed and Remix + Deno apps would import `@remix-run/deno` from some URL.
